Remove old workload
