--- hyrisecockpit/workload_generator/workloads/__init__.py
@@ -1,1 +1,0 @@
-"""Module for specific workloads."""
